feat: simple cycles faucet code redemption
