+++ e2e/assets/faucet/dfx.json
@@ -1,0 +1,8 @@
+{
+  "canisters": {
+    "faucet": {
+      "type": "motoko",
+      "main": "main.mo"
+    }
+  }
+}
