+++ e2e/assets/faucet/patch.bash
@@ -1,0 +1,1 @@
+# noop
