+++ src/dfx/src/commands/wallet/mod.rs
@@ -26,6 +26,7 @@ mod custodians;
 mod deauthorize;
 mod list_addresses;
 mod name;
+mod redeem_faucet_coupon;
 mod remove_controller;
 mod send;
 mod set_name;
@@ -45,6 +46,7 @@ enum SubCommand {
     AddController(NetworkOpts<add_controller::AddControllerOpts>),
     Authorize(NetworkOpts<authorize::AuthorizeOpts>),
     Balance(NetworkOpts<balance::WalletBalanceOpts>),
+    RedeemFaucetCoupon(NetworkOpts<redeem_faucet_coupon::RedeemFaucetCouponOpts>),
     Controllers(NetworkOpts<controllers::ControllersOpts>),
     Custodians(NetworkOpts<custodians::CustodiansOpts>),
     Deauthorize(NetworkOpts<deauthorize::DeauthorizeOpts>),
@@ -76,6 +78,9 @@ pub fn dispatch(cmd: WalletCommand) -> DfxResult {
         }
         SubCommand::Authorize(v) => with_env!(v, |env, v| authorize::exec(&env, v)),
         SubCommand::Balance(v) => with_env!(v, |env, v| balance::exec(&env, v)),
+        SubCommand::RedeemFaucetCoupon(v) => {
+            with_env!(v, |env, v| redeem_faucet_coupon::exec(&env, v))
+        }
         SubCommand::Controllers(v) => {
             with_env!(v, |env, v| controllers::exec(&env, v))
         }
